--- common/common/models/jobs/job_type.py
@@ -1,9 +1,0 @@
-# from enum import Enum
-#
-#
-## TODO: This is pretty ugly as is and maybe should be made dynamic?
-## It looks like a missing design pattern of some kind :D
-# class JobType(str, Enum):
-#    TEST = "test"
-#    HTML_MONITOR_JOB = "html_monitor_job"
-#
