+++ common/serverless_scheduler_test_plugin/models/test_job.py
@@ -1,0 +1,5 @@
+from common.models.jobs import ScheduledJob
+
+
+class TestJob(ScheduledJob):
+    __test__ = False
