+++ common/serverless_scheduler_test_plugin/plugin_export/__init__.py
@@ -1,4 +1,4 @@
-from ..models.test_job import TestJob
+from serverless_scheduler_test_plugin.models.test_job import TestJob
 
 ENUM_MAPPING = {"TEST": "test"}
 CLASS_MAPPING = {"test": TestJob}
